# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Title",
    "description": "Title",
    "version": "1.0.0"
  },
  "host": "www",
  "schemes": [
    "https"
  ],
  "paths": {
    "/users/:id":{
      "delete": {
        "summary": "Delete user",
        "description": "Delete user",
        "operationId": "deleteUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "User id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "get": {
        "summary": "Get user",
        "description": "Get user",
        "operationId": "getUser",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "User id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/": {
        "post": {
            "summary": "Create user",
            "description": "Create user",
            "operationId": "createUser",
            "produces": [
            "application/json"
            ],
            "parameters": [
            {
                "name": "name",
                "in": "body",
                "description": "User name",
                "required": true,
                "type": "string"
            },
            {
                "name": "email",
                "in": "body",
                "description": "User email",
                "required": true,
                "type": "string"
            },
            {
                "name": "password",
                "in": "body",
                "description": "User password",
                "required": true,
                "type": "string"
            },
              {
                "name": "userName",
                "in": "body",
                "description": "User name",
                "required": true,
                "type": "string"
              }
            ],
            "responses": {
            "200": {
                "description": "OK"
            },
            "400": {
                "description": "Bad Request"
            },
            "401": {
                "description": "Unauthorized"
            },
            "403": {
                "description": "Forbidden"
            },
            "404": {
                "description": "Not Found"
            },
            "500": {
                "description": "Internal Server Error"
            }
            }
        }
    },
    "/users/login": {
        "post": {
            "summary": "Login user",
            "description": "Login user",
            "operationId": "loginUser",
            "produces": [
            "application/json"
            ],
            "parameters": [
            {
                "name": "email",
                "in": "body",
                "description": "User email",
                "required": true,
                "type": "string"
            },
            {
                "name": "password",
                "in": "body",
                "description": "User password",
                "required": true,
                "type": "string"
            }
            ],
            "responses": {
            "200": {
                "description": "OK"
            },
            "400": {
                "description": "Bad Request"
            },
            "401": {
                "description": "Unauthorized"
            },
            "403": {
                "description": "Forbidden"
            },
            "404": {
                "description": "Not Found"
            },
            "500": {
                "description": "Internal Server Error"
            }
            }
        }
    }
  }
}